% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{dragon_preview}
\alias{dragon_preview}
\title{Preview mapped rows as chat turns}
\usage{
dragon_preview(dataset, n = 3)
}
\arguments{
\item{dataset}{A mapped \code{dragon_dataset}.}

\item{n}{Number of rows to show.}
}
\value{
Invisibly, a list of message lists.
}
\description{
Preview mapped rows as chat turns
}
