% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{dragon_runs_dir}
\alias{dragon_runs_dir}
\title{Directory where runs are stored}
\usage{
dragon_runs_dir()
}
\value{
A path.
}
\description{
Defaults to \code{dragonfarm_runs} under the working directory. Override with
\code{options(dragonfarm.runs_dir = ...)} or the \code{DRAGONFARM_RUNS_DIR}
environment variable.
}
